refactor: move directory location of components and layout
