+++ common/components/PostList/PostItem/index.tsx
@@ -1,12 +1,12 @@
 import Image from "next/image";
 import Link from "next/link";
 
-import Tag from "../../Tag";
+import Tag from "./Tag";
 
 import styles from "./PostItem.module.css";
 
 import type { NextPage } from "next";
-import type { PostData } from "../../../lib/posts";
+import type { PostData } from "../../../../lib/posts";
 
 interface PostItemProps {
   href: string;
